requirements.txt
aniso8601==9.0.1
apispec==6.3.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==4.4.2
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
flask-smorest==0.42.1
gunicorn==21.2.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.41.0
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper==20230918
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sympy==1.12
tiktoken==0.3.3
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
webargs==8.3.0
Werkzeug==2.3.7
